// src/app/protected/interfaces/MaterialSalida.ts
export interface MaterialSalida {

  idMs?:          number;
  idLp?:          number;
  codArticulo?:   string;
  descripcion?:   string;
  nroPaleta?:     number;
  pesoResma?:     number;
  pesoPaleta?:    number;
  pesoMaterial?:  number;
  cantidadResma?: number;
  cantidadHojas?: number;
  audUsuario?:    number;

  uuid?:          string;

}